move parts of sats.rs into input/sats.rs
Also fix any broken use path/clause.
