+++ crates/bindings-macro/src/input.rs
@@ -1,2 +1,3 @@
 pub mod table;
 pub mod util;
+pub mod sats;
